+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - MV-313 - trunck happening name if too long for tokbox
 
+### Added
+- VIMEET-2399 - Import meeting requests from CSV file
+
 ## [2.83.0]
 ### Security
 - VIMEET-2373 - fix malicious files upload